build: rename ui-components-base-theme to ui-components-theme-base
