+++ packages/@lightningjs/ui-components-test-utils/package.json
@@ -11,7 +11,7 @@
     "@lightningjs/core": "^2.9.0"
   },
   "devDependencies": {
-    "@lightningjs/ui-components-base-theme": "workspace:^",
+    "@lightningjs/ui-components-theme-base": "workspace:^",
     "jest": "^29.3.1",
     "jest-environment-jsdom": "^29.3.1",
     "jest-webgl-canvas-mock": "^0.2.3",
