+++ packages/@lightningjs/ui-components-test-utils/src/lightning-test-renderer.js
@@ -19,7 +19,7 @@
 import lng from '@lightningjs/core';
 import { updateManager, context } from '@lightningjs/ui-components';
 import jest from 'jest-mock';
-import BaseTheme from '../../ui-components-base-theme/theme';
+import BaseTheme from '../../ui-components-theme-base/theme';
 
 context.setTheme({
   ...BaseTheme,
