+++ packages/@lightningjs/ui-components-theme-base/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@lightningjs/ui-components-base-theme",
+  "name": "@lightningjs/ui-components-theme-base",
   "version": "0.0.0",
   "description": "The generic theme containing all necessary properties for the Lightning-UI shared components to map to for styling.",
   "type": "module",
