+++ packages/@lightningjs/ui-components/package.json
@@ -42,7 +42,7 @@
     "semantic-release-monorepo": "^7.0.5"
   },
   "dependencies": {
-    "@lightningjs/ui-components-base-theme": "workspace:^",
+    "@lightningjs/ui-components-theme-base": "workspace:^",
     "debounce": "^1.2.1",
     "husky": "^8.0.2",
     "lint-staged": "^13.0.3"
