+++ packages/@lightningjs/ui-components/src/components/InlineContent/InlineContent.test.js
@@ -23,7 +23,7 @@ import {
 import InlineContent from '.';
 import { getHexColor } from '../../utils';
 import { base } from './InlineContent.styles';
-import baseTheme from '@lightningjs/ui-components-base-theme';
+import baseTheme from '@lightningjs/ui-components-theme-base';
 import { context } from '../../globals/index.js';
 import { jest } from '@jest/globals';
