+++ packages/@lightningjs/ui-components/src/components/ProgressBar/ProgressBar.test.js
@@ -19,7 +19,7 @@
 import { base, tone } from './ProgressBar.styles.js';
 import { jest } from '@jest/globals';
 import { makeCreateComponent } from '@lightningjs/ui-components-test-utils';
-import baseTheme from '@lightningjs/ui-components-base-theme';
+import baseTheme from '@lightningjs/ui-components-theme-base';
 import ProgressBar from './index.js';
 
 jest.mock('./ProgressBar.styles', () => {
