+++ packages/@lightningjs/ui-components/src/globals/context/theme-manager.js
@@ -22,7 +22,7 @@ import {
   getHexColor,
   getValidColor
 } from '../../utils/index.js';
-import baseTheme from '../../../../ui-components-base-theme/theme';
+import baseTheme from '../../../../ui-components-theme-base/theme';
 import logger from './logger.js';
 import events from './events.js';
 import { fontLoader, cleanupFonts } from './fonts.js';
